Move preexisting single section workflow to separate leaf component
